Move code fromUniqueTriggerActivation to GreatPersonManager
