+++ core/src/com/unciv/ui/components/MayaCalendar.kt
@@ -2,6 +2,8 @@ package com.unciv.ui.components
 
 import com.badlogic.gdx.graphics.Color
 import com.unciv.Constants
+import com.unciv.UncivGame
+import com.unciv.logic.automation.civilization.NextTurnAutomation
 import com.unciv.logic.civilization.Civilization
 import com.unciv.models.ruleset.unique.UniqueTriggerActivation
 import com.unciv.models.ruleset.unique.UniqueType
@@ -73,12 +75,9 @@ object MayaCalendar {
         val game = civInfo.gameInfo
         val year = game.getYear()
         if (!isNewCycle(year, game.getYear(-1))) return
-        for (unique in civInfo.getMatchingUniques(UniqueType.MayanGainGreatPerson)) {
-            UniqueTriggerActivation.triggerCivwideUnique(
-                unique, civInfo,
-                notification = "{A new b'ak'tun has just begun!}\n{A Great Person joins you!}"
-            )
-        }
+        civInfo.greatPeople.triggerMayanGreatPerson()
+        if (civInfo.isAI() || UncivGame.Current.settings.autoPlay.isAutoPlayingAndFullAI())
+            NextTurnAutomation.chooseGreatPerson(civInfo)
     }
 
     // User interface to explain changed year display
